requirements.txt
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
transformers==4.40.0
torch==2.2.2
spacy==3.7.4
nltk==3.8.1
keybert==0.8.4
sentence-transformers==2.7.0
streamlit==1.33.0
streamlit-extras==0.4.2
wordcloud==1.9.3
matplotlib==3.8.4
plotly==5.20.0
pandas==2.2.2
python-dotenv==1.0.1
loguru==0.7.2